Peach_Planet/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
Pillow==11.3.0
